Splitting code apart, rm'ing now unneeded device actions/reducer
