+++ lib/windows/app/containers/DeviceSelectorContainer.js
@@ -34,32 +34,157 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import DeviceLister from 'nrf-device-lister';
+
+import React from 'react';
+import PropTypes from 'prop-types';
+
 import DeviceSelector from '../components/DeviceSelector';
-import * as DeviceActions from '../actions/deviceActions';
-import withHotkey from '../../../util/withHotkey';
 import { connect } from '../../../util/apps';
+import { logger } from '../../../api/logging';
 
-function mapStateToProps(state) {
-    const { device } = state.core;
+/*
+ * Stateful device selector.
+ *
+ * Displays a drop-down list of nRF devices so the user can choose one of
+ * them.
+ *
+ * This uses nrf-device-lister under the hood. Depending on what kind
+ * of devices the application needs, developers can specify a `capabilities`
+ * object (e.g. look only for serial ports, look only for j-link probes, look
+ * for several things at once). This will be passed to nrf-device-lister.
+ *
+ * Interaction-wise, this component does two things and two things *only*:
+ * - Dispatch a DEVICE_SELECTED action whenever the user selects
+ *     a device from the drop-down list. This action has a `device` property
+ *     containing the data structure for the device as given by
+ *     nrf-device-lister.
+ * - Dispatch a DEVICE_DESELECTED action whenever the user selects the
+ *     "close device" option from the drop-down list, or whenever the
+ *     user selects a different device.
+ *
+ * This file implements the logic. The view/rendering/template/whatchacallit
+ * stays in ../components/DeviceSelector
+ */
 
-    return {
-        devices: device.devices,
-        selectedDevice: device.selectedDevice,
-        isExpanded: device.isSelectorExpanded,
-        isLoading: device.isLoading,
-    };
+class DeviceSelectorContainer extends React.Component {
+    constructor(props) {
+        super(props);
+
+        this.deviceLister = new DeviceLister(props.capabilities);
+
+        this.state = {
+            devices: [],
+            selectedSerialNumber: undefined,
+        };
+
+        this.deviceLister.on('conflated', devices => {
+            this.setState(prev =>
+//    logger.info(`Changes in available devices. Now ${Array.from(devices).length} devices.`);
+                 ({ ...prev, devices }));
+        });
+
+        this.deviceLister.on('error', err => {
+            if (err.usb) {
+                const usbAddr = `${err.usb.device.busNumber}.${err.usb.device.deviceAddress}`;
+
+                let errMsg = (`Error while probing usb device at bus.address ${usbAddr}. `);
+                if (process.platform === 'linux') {
+                    errMsg += 'Please check your udev rules concerning permissions for USB devices.';
+                } else if (process.platform === 'win32') {
+                    errMsg += 'Please check that a libusb-compatible kernel driver is bound to this device.';
+                }
+
+                logger.error(err.error.toString());
+                logger.error(errMsg);
+            } else {
+                logger.error(`Error while probing devices: ${err.error}`);
+            }
+        });
+
+        this.deviceLister.start();
+    }
+
+    /*
+     * Called from the templated selector.
+     * Shall receive a device definition.
+     * Resets the internal state, and calls the downstream onSelect().
+     * Note that this.onSelect() is different from this.props.onSelect(),
+     * the later one is the one that actually triggers the action.
+     */
+    onSelect(device) {
+        if (this.state.selectDevice) {
+            this.onDeselect();
+        }
+//                 logger.info(`Selecting device with s/n ${device.serialNumber}`);
+        this.setState(prev => ({ ...prev, selectedSerialNumber: device.serialNumber }));
+        this.props.onSelect(device);
+    }
+
+    /*
+     * Called from the templated selector.
+     * Resets the internal state, and calls the downstream onDeselect().
+     * Note that this.onDeselect() is different from this.props.onDeselect(),
+     * the later one is the one that actually triggers the action.
+     */
+    onDeselect() {
+//                 logger.info('Deselecting device');
+        this.setState(prev => ({ ...prev, selectedSerialNumber: undefined }));
+        this.props.onDeselect();
+    }
+
+
+    /*
+     * Returns the JSX corresponding to a drop-down menu.
+     * Prepares some properties for the template, uses it.
+     */
+    render() {
+        const togglerText = this.state.selectedSerialNumber ?
+            this.state.selectedSerialNumber.toString() :
+            'Select device';
+
+        const displayCloseItem = Boolean(this.state.selectedSerialNumber);
+        const devices = Array.from(this.state.devices);
+
+        const templateProps = {
+            ...this.props,
+            togglerText,
+            displayCloseItem,
+            devices,
+            onSelect: this.onSelect,
+            onDeselect: this.onDeselect,
+        };
+
+        return (
+            <DeviceSelector {...templateProps} />
+        );
+    }
 }
+
+DeviceSelectorContainer.propTypes = {
+    onSelect: PropTypes.func.isRequired,
+    onDeselect: PropTypes.func.isRequired,
+    capabilities: PropTypes.shape({}),
+};
+
+DeviceSelectorContainer.defaultProps = {
+    capabilities: undefined,
+};
 
 function mapDispatchToProps(dispatch) {
     return {
-        onSelect: (device, verifySerialPortAvailable) =>
-            dispatch(DeviceActions.selectDevice(device, verifySerialPortAvailable)),
-        onDeselect: () => dispatch(DeviceActions.deselectDevice()),
-        toggleExpanded: () => dispatch(DeviceActions.toggleSelectorExpanded()),
+        onSelect: (device) =>
+            dispatch({
+                type: 'DEVICE_SELECTED',
+                device
+            }),
+        onDeselect: () => dispatch({
+            type: 'DEVICE_DESELECTED'
+        }),
     };
 }
 
 export default connect(
-    mapStateToProps,
+    args => args,
     mapDispatchToProps,
-)(withHotkey(DeviceSelector), 'DeviceSelector');
+)(DeviceSelectorContainer, 'DeviceSelector');
