+++ lib/windows/app/reducers/coreReducer.js
@@ -37,7 +37,6 @@
 import { combineReducers } from 'redux';
 import navMenu from './navMenuReducer';
 import log from './logReducer';
-import device from './deviceReducer';
 import serialPort from './serialPortReducer';
 import firmwareDialog from './firmwareDialogReducer';
 import errorDialog from '../../../reducers/errorDialogReducer';
@@ -45,7 +44,6 @@ import errorDialog from '../../../reducers/errorDialogReducer';
 export default combineReducers({
     navMenu,
     log,
-    device,
     serialPort,
     firmwareDialog,
     errorDialog,
